Fixed CI errors
