+++ ballista/executor/src/config.rs
@@ -26,11 +26,10 @@
 // as mentioned in https://github.com/apache/datafusion-ballista/issues/1271
 //
 #![allow(clippy::uninlined_format_args, clippy::unused_unit)]
-use std::fmt::Display;
-
 use ballista_core::error::BallistaError;
 
 use crate::executor_process::ExecutorProcessConfig;
+use crate::metrics::ExecutorMetricCollectionPolicy;
 
 /// Command-line arguments for configuring a Ballista executor.
 ///
@@ -189,40 +188,5 @@ impl TryFrom<Config> for ExecutorProcessConfig {
             override_arrow_flight_service: None,
             override_create_grpc_client_endpoint: None,
         })
-    }
-}
-
-/// Configures which executor's metrics should be collected
-#[derive(Clone, Copy, Debug, serde::Deserialize, Default)]
-#[cfg_attr(feature = "build-binary", derive(clap::ValueEnum))]
-pub enum ExecutorMetricCollectionPolicy {
-    /// Collect only system-wide metrics
-    SystemOnly,
-    /// Collect only current process metrics
-    ProcessOnly,
-    /// Collect both system-wide and process metrics
-    SystemAndProcess,
-    /// Default value - no metrics collected
-    #[default]
-    Off,
-}
-
-impl Display for ExecutorMetricCollectionPolicy {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ExecutorMetricCollectionPolicy::SystemOnly => f.write_str("sys"),
-            ExecutorMetricCollectionPolicy::ProcessOnly => f.write_str("proc"),
-            ExecutorMetricCollectionPolicy::SystemAndProcess => f.write_str("all"),
-            ExecutorMetricCollectionPolicy::Off => f.write_str("off"),
-        }
-    }
-}
-
-#[cfg(feature = "build-binary")]
-impl std::str::FromStr for ExecutorMetricCollectionPolicy {
-    type Err = String;
-
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        clap::ValueEnum::from_str(s, true)
     }
 }
