+++ ballista/executor/src/metrics/mod.rs
@@ -17,7 +17,7 @@
 
 use crate::execution_engine::QueryStageExecutor;
 use log::info;
-use std::sync::Arc;
+use std::{fmt::Display, sync::Arc};
 
 /// `ExecutorMetricsCollector` records metrics for `ShuffleWriteExec`
 /// after they are executed.
@@ -51,5 +51,40 @@ impl ExecutorMetricsCollector for LoggingMetricsCollector {
         info!(
             "\n=== [{job_id}/{stage_id}/{partition}] Physical plan with metrics ===\n{plan}\n"
         );
+    }
+}
+
+/// Configures which executor's metrics should be collected
+#[derive(Clone, Copy, Debug, serde::Deserialize, Default)]
+#[cfg_attr(feature = "build-binary", derive(clap::ValueEnum))]
+pub enum ExecutorMetricCollectionPolicy {
+    /// Collect only system-wide metrics
+    SystemOnly,
+    /// Collect only current process metrics
+    ProcessOnly,
+    /// Collect both system-wide and process metrics
+    SystemAndProcess,
+    /// Default value - no metrics collected
+    #[default]
+    Off,
+}
+
+impl Display for ExecutorMetricCollectionPolicy {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ExecutorMetricCollectionPolicy::SystemOnly => f.write_str("sys"),
+            ExecutorMetricCollectionPolicy::ProcessOnly => f.write_str("proc"),
+            ExecutorMetricCollectionPolicy::SystemAndProcess => f.write_str("all"),
+            ExecutorMetricCollectionPolicy::Off => f.write_str("off"),
+        }
+    }
+}
+
+#[cfg(feature = "build-binary")]
+impl std::str::FromStr for ExecutorMetricCollectionPolicy {
+    type Err = String;
+
+    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
+        clap::ValueEnum::from_str(s, true)
     }
 }
